refactor: S3 Util 사용할 수 있도록 수정 (DURI-349)
