+++ app/src/main/java/kr/com/duri/common/s3/S3Util.java
@@ -1,8 +1,10 @@
 package kr.com.duri.common.s3;
 
 import java.io.IOException;
+import java.time.LocalDateTime;
 import java.util.UUID;
 
+import kr.com.duri.user.domain.entity.ReviewImage;
 import kr.com.duri.user.exception.ReviewImageUploadException;
 import lombok.RequiredArgsConstructor;
 
@@ -24,17 +26,17 @@ public class S3Util {
 
     private final AmazonS3 amazonS3;
     private static final String S3_BASIC_URL = "https://%s.s3.%s.amazonaws.com/%s";
-    private static final String S3_FOLDER_NAME = "review";
 
-    public String uploadToS3(MultipartFile image) {
+    // S3 사진 업로드
+    public String uploadToS3(MultipartFile image, String S3_folder_name) {
         try {
             // 고유 파일명
             String fileName =
-                    S3_FOLDER_NAME
+                    S3_folder_name
                             + "/"
                             + UUID.randomUUID()
                             + "_"
-                            + image.getOriginalFilename(); // 폴더 내부에 랜덤숫자_원본파일명
+                            + LocalDateTime.now(); // 폴더 내부에 랜덤숫자_원본파일명
             // 메타 데이터
             ObjectMetadata metadata = new ObjectMetadata();
             metadata.setContentType(image.getContentType());
@@ -45,14 +47,22 @@ public class S3Util {
             amazonS3.putObject(putObjectRequest);
             return String.format(S3_BASIC_URL, bucket, amazonS3.getRegionName(), fileName);
         } catch (IOException e) {
-            throw new ReviewImageUploadException("해당 이미지를 불러오는 데 실패하였습니다.");
+            throw new ReviewImageUploadException("이미지 로딩 실패");
         } catch (Exception e) {
-            throw new ReviewImageUploadException("해당 이미지를 S3에 업로드하는 데 실패하였습니다.");
+            throw new ReviewImageUploadException("이미지 S3 업로드 실패");
         }
     }
 
-    public void deleteFromS3(String imageUrl) {
-        String deleteS3Key = imageUrl.substring(imageUrl.indexOf(".com/") + 5);
+    // S3 삭제
+    public void deleteFromS3(String imageURL) {
+        String deleteS3Key = imageURL.substring(imageURL.indexOf(".com/") + 5);
+        amazonS3.deleteObject(new DeleteObjectRequest(bucket, deleteS3Key));
+    }
+
+    // S3 삭제
+    public void deleteFromS3ByReviewImage(ReviewImage reviewImage) {
+        String originS3Url = reviewImage.getImage();
+        String deleteS3Key = originS3Url.substring(originS3Url.indexOf(".com/") + 5);
         amazonS3.deleteObject(new DeleteObjectRequest(bucket, deleteS3Key));
     }
 }
