+++ app/src/main/java/kr/com/duri/user/application/service/ReviewImageService.java
@@ -1,6 +1,5 @@
 package kr.com.duri.user.application.service;
 
-import java.io.IOException;
 import java.util.List;
 
 import kr.com.duri.user.domain.entity.Review;
@@ -18,12 +17,6 @@ public interface ReviewImageService {
 
     // 추가, 수정
     Boolean saveReviewImage(MultipartFile image, Review review);
-
-    // S3 업로드
-    String uploadToS3(MultipartFile image) throws IOException;
-
-    // S3 삭제
-    void deleteFromS3(ReviewImage reviewImage);
 
     // 삭제
     Boolean deleteReview(Long reviewId);
