+++ app/src/main/java/kr/com/duri/user/application/service/impl/AiServiceimpl.java
@@ -1,12 +1,9 @@
 package kr.com.duri.user.application.service.impl;
 
 import java.io.IOException;
-import java.time.LocalDateTime;
 import java.util.Map;
-import java.util.UUID;
 
 import kr.com.duri.user.application.service.AiService;
-import kr.com.duri.user.exception.ReviewImageUploadException;
 import lombok.RequiredArgsConstructor;
 
 import org.apache.hc.client5.http.classic.methods.HttpGet;
@@ -18,25 +15,13 @@ import org.apache.hc.core5.http.HttpStatus;
 import org.apache.hc.core5.http.io.entity.StringEntity;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.DeleteObjectRequest;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Service
 @RequiredArgsConstructor
 public class AiServiceimpl implements AiService {
-
-    @Value("${spring.cloud.aws.s3.bucket}")
-    private String bucket;
-
-    private final AmazonS3 amazonS3;
-    private static final String S3_BASIC_URL = "https://%s.s3.%s.amazonaws.com/%s";
-    private static final String S3_FOLDER_NAME = "ai";
 
     @Value("${replicate.api.token}")
     private String replicateApiKey;
@@ -73,33 +58,6 @@ For a "Lion cut", cut short the body fur, leaving the chest and neck fur longer.
                     "베이비", "c7c4a7c1ed0baa97a4e9f12b76fa2377aad131504afd405b10abb91d0f1af899",
                     "라이언", "c7c4a7c1ed0baa97a4e9f12b76fa2377aad131504afd405b10abb91d0f1af899");
 
-    // S3 사진 업로드
-    public String uploadS3(MultipartFile image) {
-        try {
-            // 고유 파일명
-            String fileName =
-                    S3_FOLDER_NAME
-                            + "/"
-                            + UUID.randomUUID()
-                            + "_"
-                            + LocalDateTime.now(); // 폴더 내부에 랜덤숫자_원본파일명
-            // 메타 데이터
-            ObjectMetadata metadata = new ObjectMetadata();
-            metadata.setContentType(image.getContentType());
-            metadata.setContentLength(image.getSize());
-            // S3에 파일 업로드
-            PutObjectRequest putObjectRequest =
-                    new PutObjectRequest(bucket, fileName, image.getInputStream(), metadata);
-            amazonS3.putObject(putObjectRequest);
-            System.out.println(" ");
-            return String.format(S3_BASIC_URL, bucket, amazonS3.getRegionName(), fileName);
-        } catch (IOException e) {
-            throw new ReviewImageUploadException("이미지 로딩 실패");
-        } catch (Exception e) {
-            throw new ReviewImageUploadException("이미지 S3 업로드 실패");
-        }
-    }
-
     // 프롬프트 생성
     public String generatePrompt(String styleText) {
         String stylePrompt = STYLE_PROMPTS.get(styleText);
@@ -109,12 +67,12 @@ For a "Lion cut", cut short the body fur, leaving the chest and neck fur longer.
         return """
         - Short and neat fur on the dog's body. Apply clean trimming to all visible fur. Fur length no longer than 0.5 cm.
         - Apply a specific grooming style to dog's coat.
-    """
+        """
                 + stylePrompt
                 + """
         - Make sure the final image clearly highlights your intended grooming style.
         - At this time, the dog's face and overall structure must remain natural and realistic.
-    """;
+        """;
     }
 
     // 네거티브 프롬프트 생성
@@ -136,22 +94,17 @@ For a "Lion cut", cut short the body fur, leaving the chest and neck fur longer.
     public String callReplicateApi(
             String imageUrl, String prompt, String negativePrompt, String version)
             throws IOException {
-        System.out.println("model : " + version);
-        System.out.println("S3 업로드된 주소 : " + imageUrl);
-        System.out.println("propmt : " + prompt);
-        System.out.println("negative : " + negativePrompt);
 
-        // HTTP 클라이언트를 사용하여 Replicate API 호출
+        // 1) HTTP 클라이언트를 사용하여 Replicate API 호출
         CloseableHttpClient httpClient = HttpClients.createDefault();
         HttpPost request = new HttpPost("https://api.replicate.com/v1/predictions");
-        // API에게 보낼 Request Body (Prompt Hyper Parameter) 설정
+        // 2) API에게 보낼 Request Body (Prompt Hyper Parameter) 설정
         String json =
                 new ObjectMapper()
                         .writeValueAsString(
                                 Map.of(
                                         "version",
                                         version,
-                                        // "c7c4a7c1ed0baa97a4e9f12b76fa2377aad131504afd405b10abb91d0f1af899",
                                         "input",
                                         Map.of(
                                                 "image",
@@ -176,19 +129,19 @@ For a "Lion cut", cut short the body fur, leaving the chest and neck fur longer.
                                                 55)));
         StringEntity entity = new StringEntity(json);
         request.setEntity(entity);
-        // 헤더 추가
+        // 3) 헤더 추가
         request.setHeader("Authorization", "Bearer " + replicateApiKey);
         request.setHeader("Content-Type", "application/json");
         request.setHeader("Prefer", "wait");
-        // 요청 및 응답 확인
+        // 4) 요청 및 응답
         try (CloseableHttpResponse response = httpClient.execute(request)) {
             if (response.getCode() == HttpStatus.SC_ACCEPTED
                     || response.getCode() == HttpStatus.SC_CREATED) {
-                // API 응답에서 결과 이미지 URL 파싱
+                // 5) API 응답에서 결과 이미지 URL 파싱
                 JsonNode responseBody =
                         new ObjectMapper().readTree(response.getEntity().getContent());
                 String statusUrl = responseBody.get("urls").get("get").asText();
-                // 계속 상태 확인
+                // 6) 계속 상태 확인
                 while (true) {
                     HttpGet statusRequest = new HttpGet(statusUrl);
                     statusRequest.setHeader("Authorization", "Bearer " + replicateApiKey);
@@ -201,16 +154,16 @@ For a "Lion cut", cut short the body fur, leaving the chest and neck fur longer.
                         String body = statusBody.asText();
 
                         if ("succeeded".equals(status)) {
-                            // 작업 성공: 결과 URL 반환
+                            // 7) 작업 성공: 결과 URL 반환
                             return statusBody.get("output").get(0).asText();
                         } else if ("failed".equals(status)) {
-                            // 작업 실패 처리
+                            // 8) 작업 실패 처리
                             throw new RuntimeException(
                                     "Replicate API 응답 실패:"
                                             + statusBody.get("error").asText()
                                             + body);
                         }
-                        // 일정 시간 대기 후 다시 조회
+                        // 9) 일정 시간 대기 후 다시 조회
                         Thread.sleep(2000);
                     }
                 }
@@ -220,11 +173,5 @@ For a "Lion cut", cut short the body fur, leaving the chest and neck fur longer.
         } catch (InterruptedException e) {
             throw new RuntimeException("상태 확인 중 인터럽트 발생");
         }
-    }
-
-    // S3 삭제
-    public void deleteFromS3(String imageURL) {
-        String deleteS3Key = imageURL.substring(imageURL.indexOf(".com/") + 5);
-        amazonS3.deleteObject(new DeleteObjectRequest(bucket, deleteS3Key));
     }
 }
