+++ app/src/main/java/kr/com/duri/user/application/service/impl/SiteUserServiceImpl.java
@@ -73,7 +73,7 @@ public class SiteUserServiceImpl implements SiteUserService {
         if (img == null || img.isEmpty()) {
             return null;
         }
-        return s3Util.uploadToS3(img);
+        return s3Util.uploadToS3(img, "info");
     }
 
     @Override
